Add files via upload
This is a library management system created using python, csv and tkinter. The tkinter module is used to create a GUI for the library's website. It is used to sign up, login, access accounts, browse through available books, display the borrowing and returning of books, display overdue books and the date of borrowing. The csv files are used to store account info such as usernames and passwords, and to keep records of which books have been borrowed, returned or are overdue in the name of the user.